+++ lib/flutter_dothantech_lpapi_thermal_printer_platform_interface.dart
@@ -56,11 +56,12 @@ abstract class LpapiThermalPrinterPlatform extends PlatformInterface {
     throw UnimplementedError('printText() has not been implemented.');
   }
 
+  /// Print 1D barcode product label (50x40mm landscape)
   Future<bool> print1DBarcode(
     String barcode, {
     String text = '',
-    int width = 48,
-    int height = 50,
+    int width = 50,
+    int height = 40,
   }) {
     throw UnimplementedError('print1DBarcode() has not been implemented.');
   }